# apps/router/main.go
package main

import (
	"bytes"
	"context"
	"crypto/sha256"
	"crypto/subtle"
	"fmt"
	"io"
	"log/slog"
	"net/http"
	"net/http/httputil"
	"net/url"
	"os"
	"os/signal"
	"regexp"
	"strings"
	"syscall"
	"time"

	"cloudmeter/internal/config"
	"github.com/jackc/pgx/v5"
	"github.com/jackc/pgx/v5/pgxpool"
)

const appAccessCookie = "cloudmeter_app_access"

func main() {
	if len(os.Args) > 1 && os.Args[1] == "healthcheck" {
		response, err := (&http.Client{Timeout: 2 * time.Second}).Get("http://127.0.0.1:8082/healthz")
		if err != nil || response.StatusCode != http.StatusOK {
			os.Exit(1)
		}
		_ = response.Body.Close()
		return
	}
	logger := slog.New(slog.NewJSONHandler(os.Stdout, nil))
	cfg, err := config.Load()
	if err != nil {
		logger.Error("invalid configuration", "error", err)
		os.Exit(1)
	}
	cfg.RouterToken = strings.TrimSpace(cfg.RouterToken)
	if len(cfg.RouterToken) < 32 {
		logger.Error("ROUTER_INTERNAL_TOKEN must contain at least 32 characters")
		os.Exit(1)
	}
	ctx, stop := signal.NotifyContext(context.Background(), syscall.SIGINT, syscall.SIGTERM)
	defer stop()
	db, err := pgxpool.New(ctx, cfg.DatabaseURL)
	if err != nil {
		logger.Error("database configuration failed", "error", err)
		os.Exit(1)
	}
	defer db.Close()
	mux := http.NewServeMux()
	mux.HandleFunc("GET /healthz", func(w http.ResponseWriter, _ *http.Request) {
		w.Header().Set("Content-Type", "application/json")
		_, _ = w.Write([]byte(`{"status":"ok"}`))
	})
	mux.Handle("/apps/", requireRouterToken(cfg.RouterToken, routeHandler(db, logger)))
	server := &http.Server{Addr: ":8082", Handler: mux, ReadHeaderTimeout: 5 * time.Second, IdleTimeout: 60 * time.Second}
	go func() {
		logger.Info("app router listening", "addr", server.Addr)
		if err := server.ListenAndServe(); err != nil && err != http.ErrServerClosed {
			logger.Error("router stopped", "error", err)
			stop()
		}
	}()
	<-ctx.Done()
	shutdown, cancel := context.WithTimeout(context.Background(), 10*time.Second)
	defer cancel()
	_ = server.Shutdown(shutdown)
}

func requireRouterToken(token string, next http.Handler) http.Handler {
	return http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		provided := r.Header.Get("X-CloudMeter-Router-Token")
		if len(provided) != len(token) || subtle.ConstantTimeCompare([]byte(provided), []byte(token)) != 1 {
			http.Error(w, "forbidden", http.StatusForbidden)
			return
		}
		next.ServeHTTP(w, r)
	})
}

func routeHandler(db *pgxpool.Pool, logger *slog.Logger) http.Handler {
	return http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		cookie, cookieErr := r.Cookie(appAccessCookie)
		if cookieErr != nil || strings.TrimSpace(cookie.Value) == "" {
			http.Error(w, "application sign-in required", http.StatusUnauthorized)
			return
		}
		tokenHash := sha256.Sum256([]byte(cookie.Value))
		var sessionUserID string
		if err := db.QueryRow(r.Context(), `SELECT session.user_id::text FROM sessions session
			JOIN users ON users.id=session.user_id AND users.status='active'
			WHERE session.token_hash=$1 AND session.revoked_at IS NULL AND session.expires_at>now()`, tokenHash[:]).Scan(&sessionUserID); err != nil {
			if err == pgx.ErrNoRows {
				http.Error(w, "application session is invalid or expired", http.StatusUnauthorized)
				return
			}
			logger.Error("application session lookup failed", "error", err)
			http.Error(w, "session lookup failed", http.StatusBadGateway)
			return
		}
		var prefix, host string
		var port int
		var routeSpec map[string]any
		err := db.QueryRow(r.Context(), `SELECT
		'/apps/'||u.slug||'/'||a.slug,
		'release-'||left(replace(rel.id::text,'-',''),12),
		coalesce(nullif(rel.immutable_snapshot->'route_spec'->>'port','')::int,nullif(rel.immutable_snapshot->'route_spec'->>'containerPort','')::int,8080),
		coalesce(rel.immutable_snapshot->'route_spec','{}'::jsonb)
		FROM app_routes ar
		JOIN user_apps a ON a.id=ar.user_app_id AND a.instance_id=ar.instance_id
		JOIN users u ON u.id=a.user_id
		JOIN app_releases rel ON rel.id=ar.release_id AND rel.user_app_id=a.id
		WHERE a.user_id=$2 AND u.status='active' AND a.status IN ('running','updating') AND a.last_successful_release_id=rel.id AND rel.state='active'
		  AND ($1='/apps/'||u.slug||'/'||a.slug OR $1 LIKE '/apps/'||u.slug||'/'||a.slug||'/%')
		ORDER BY length('/apps/'||u.slug||'/'||a.slug) DESC LIMIT 1`, r.URL.Path, sessionUserID).Scan(&prefix, &host, &port, &routeSpec)
		if err == pgx.ErrNoRows {
			http.NotFound(w, r)
			return
		}
		if err != nil {
			logger.Error("route lookup failed", "error", err)
			http.Error(w, "route lookup failed", http.StatusBadGateway)
			return
		}
		if redirectAppRoot(w, r, prefix) {
			return
		}
		if isWebSocketRequest(r) && !routeBoolean(routeSpec, "websocket", true) {
			http.Error(w, "websocket is disabled for this application", http.StatusUpgradeRequired)
			return
		}
		if strings.Contains(strings.ToLower(r.Header.Get("Accept")), "text/event-stream") && !routeBoolean(routeSpec, "sse", true) {
			http.Error(w, "server-sent events are disabled for this application", http.StatusNotAcceptable)
			return
		}
		target, _ := url.Parse(fmt.Sprintf("http://%s:%d", host, port))
		proxy := httputil.NewSingleHostReverseProxy(target)
		rootRequest := r.Method == http.MethodGet && r.URL.Path == prefix+"/" && !isWebSocketRequest(r)
		if routeBoolean(routeSpec, "sse", true) {
			proxy.FlushInterval = -1
		}
		original := proxy.Director
		proxy.Director = func(req *http.Request) {
			original(req)
			removeRequestCookie(req, appAccessCookie)
			if routeBoolean(routeSpec, "stripPrefix", true) {
				path := strings.TrimPrefix(req.URL.Path, prefix)
				if path == "" {
					path = "/"
				}
				req.URL.Path = joinURLPath(routeString(routeSpec, "basePath", "/"), path)
			}
			if rootRequest {
				// Let Go transparently decode an upstream gzip response so the
				// public-path adapter can safely inspect the root HTML document.
				req.Header.Del("Accept-Encoding")
				// A 304 has no body to adapt. Always fetch the root document so a
				// previously cached unprefixed <base> cannot bypass the adapter.
				req.Header.Del("If-None-Match")
				req.Header.Del("If-Modified-Since")
			}
			req.Header.Set("X-Forwarded-Prefix", prefix)
		}
		cookiePath := routeString(routeSpec, "cookiePath", "")
		proxy.ModifyResponse = func(response *http.Response) error {
			stripReservedResponseCookie(response.Header, appAccessCookie)
			if cookiePath != "" || rootRequest {
				if rootRequest {
					if err := rewriteRootHTMLBase(response, prefix); err != nil {
						return err
					}
				}
				if cookiePath != "" {
					values := response.Header.Values("Set-Cookie")
					if len(values) > 0 {
						response.Header.Del("Set-Cookie")
						for _, value := range values {
							response.Header.Add("Set-Cookie", rewriteCookiePath(value, cookiePath, prefix))
						}
					}
				}
			}
			return nil
		}
		proxy.ErrorHandler = func(w http.ResponseWriter, _ *http.Request, err error) {
			logger.Error("upstream failed", "host", host, "error", err)
			http.Error(w, "application unavailable", http.StatusBadGateway)
		}
		proxy.ServeHTTP(w, r)
	})
}

func removeRequestCookie(r *http.Request, reserved string) {
	values := make([]string, 0)
	for _, cookie := range r.Cookies() {
		if cookie.Name != reserved {
			values = append(values, cookie.Name+"="+cookie.Value)
		}
	}
	if len(values) == 0 {
		r.Header.Del("Cookie")
	} else {
		r.Header.Set("Cookie", strings.Join(values, "; "))
	}
}

func stripReservedResponseCookie(header http.Header, reserved string) {
	values := header.Values("Set-Cookie")
	if len(values) == 0 {
		return
	}
	header.Del("Set-Cookie")
	for _, value := range values {
		name, _, _ := strings.Cut(value, "=")
		if !strings.EqualFold(strings.TrimSpace(name), reserved) {
			header.Add("Set-Cookie", value)
		}
	}
}

const maxRootHTMLRewriteBytes = 4 << 20

var rootHTMLBasePattern = regexp.MustCompile(`(?i)(<base\b[^>]*\bhref\s*=\s*["'])/(["'])`)

type joinedReadCloser struct {
	io.Reader
	io.Closer
}

func rewriteRootHTMLBase(response *http.Response, publicPath string) error {
	if response.StatusCode < http.StatusOK || response.StatusCode >= http.StatusMultipleChoices ||
		!strings.HasPrefix(strings.ToLower(response.Header.Get("Content-Type")), "text/html") ||
		(response.Header.Get("Content-Encoding") != "" && !strings.EqualFold(response.Header.Get("Content-Encoding"), "identity")) {
		return nil
	}
	body := response.Body
	payload, err := io.ReadAll(io.LimitReader(body, maxRootHTMLRewriteBytes+1))
	if err != nil {
		return err
	}
	if len(payload) > maxRootHTMLRewriteBytes {
		response.Body = &joinedReadCloser{Reader: io.MultiReader(bytes.NewReader(payload), body), Closer: body}
		return nil
	}
	_ = body.Close()
	prefix := strings.TrimRight(publicPath, "/") + "/"
	rewritten := rootHTMLBasePattern.ReplaceAll(payload, []byte("$1"+prefix+"$2"))
	response.Body = io.NopCloser(bytes.NewReader(rewritten))
	response.ContentLength = int64(len(rewritten))
	response.Header.Set("Content-Length", fmt.Sprint(len(rewritten)))
	if !bytes.Equal(payload, rewritten) {
		response.Header.Del("ETag")
		response.Header.Del("Content-MD5")
	}
	return nil
}

func redirectAppRoot(w http.ResponseWriter, r *http.Request, prefix string) bool {
	if r.URL.Path != prefix || isWebSocketRequest(r) {
		return false
	}
	location := prefix + "/"
	if r.URL.RawQuery != "" {
		location += "?" + r.URL.RawQuery
	}
	http.Redirect(w, r, location, http.StatusPermanentRedirect)
	return true
}

func routeBoolean(spec map[string]any, key string, fallback bool) bool {
	value, ok := spec[key].(bool)
	if !ok {
		return fallback
	}
	return value
}

func routeString(spec map[string]any, key, fallback string) string {
	value, ok := spec[key].(string)
	if !ok || value == "" {
		return fallback
	}
	return value
}

func joinURLPath(base, suffix string) string {
	if base == "" || base == "/" {
		if suffix == "" {
			return "/"
		}
		if strings.HasPrefix(suffix, "/") {
			return suffix
		}
		return "/" + suffix
	}
	if suffix == "" || suffix == "/" {
		return strings.TrimRight(base, "/") + "/"
	}
	return strings.TrimRight(base, "/") + "/" + strings.TrimLeft(suffix, "/")
}

func isWebSocketRequest(r *http.Request) bool {
	return strings.EqualFold(strings.TrimSpace(r.Header.Get("Upgrade")), "websocket") && strings.Contains(strings.ToLower(r.Header.Get("Connection")), "upgrade")
}

func rewriteCookiePath(header, upstreamPath, publicPath string) string {
	parts := strings.Split(header, ";")
	found := false
	for index := 1; index < len(parts); index++ {
		attribute := strings.TrimSpace(parts[index])
		name, value, ok := strings.Cut(attribute, "=")
		if !ok || !strings.EqualFold(strings.TrimSpace(name), "Path") {
			continue
		}
		found = true
		if strings.TrimSpace(value) == upstreamPath {
			parts[index] = " Path=" + publicPath
		}
	}
	if !found && upstreamPath == "/" {
		parts = append(parts, " Path="+publicPath)
	}
	return strings.Join(parts, ";")
}