# docs/openapi.yaml
openapi: 3.1.0
info:
  title: CloudMeter API
  version: 0.1.0
servers:
  - url: /api
paths:
...
  /apps:
    get:
      operationId: listApps
      security: [{bearerAuth: []}]
      responses:
        '200': {description: "User applications including active publicPath, configured CPU and memory limits, latest point-in-time Docker CPU and memory usage, and metricsSampledAt. Runtime metrics are omitted until a fresh sample is available."}
    post:
      operationId: createApp
      security: [{bearerAuth: []}]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [productId, versionId, slug, idempotencyKey]
              properties:
                productId: {type: string, format: uuid}
                versionId: {type: string, format: uuid}
                slug: {type: string, pattern: '^[a-z0-9][a-z0-9-]{0,62}$'}
                idempotencyKey: {type: string, minLength: 1, maxLength: 128}
                secrets:
                  type: object
                  additionalProperties: {type: string, minLength: 1, maxLength: 65536}
                  description: Values for every Secret key declared by the selected product version. Values are encrypted and never returned.
                resources:
                  type: object
                  description: User-selected overrides. The API accepts only fields explicitly enabled by the immutable product version.
                  properties:
                    cpuCores: {type: number, minimum: 0.1, maximum: 64}
                    memoryMiB: {type: number, minimum: 64, maximum: 262144}
                    dataVolumeGiB: {type: number, minimum: 1, maximum: 16384, description: One capacity shared by every persistent mount in this application and billed once.}
                    command: {type: array, items: {type: string}}
                    environment: {type: object, additionalProperties: {type: string}}
                    dependencies: {type: array, items: {$ref: '#/components/schemas/ProductDependency'}}
      responses:
        '201': {description: "Application, encrypted initial Secrets, immutable release, and deployment job created atomically"}
        '400': {description: "Invalid resources, slug, or missing or undeclared initial Secrets"}
        '403': {description: Product is not included in the current plan}
        '409': {description: "Subscription, application, deployment concurrency, CPU, memory, system disk, or data disk quota exceeded; or a required product dependency is unavailable"}
  /apps/{appID}:
    delete:
      operationId: deleteApp
      description: Soft-deletes an owned application, removes its route, stops billing, and queues container and persistent-volume cleanup while retaining financial and release history.
      security: [{bearerAuth: []}]
      parameters:
        - {name: appID, in: path, required: true, schema: {type: string, format: uuid}}
      responses:
        '202': {description: Application hidden and resource cleanup queued}
        '404': {description: Application not found}
        '409': {description: An operation is active or another application depends on this service}
...
  /apps/{appID}/route:
    get:
      operationId: appRoute
      security: [{bearerAuth: []}]
      parameters:
        - {name: appID, in: path, required: true, schema: {type: string, format: uuid}}
      responses:
        '200': {description: Active application route}
        '404': {description: Route not active}
...
  /admin/products/{productID}:
    delete:
      operationId: deleteProduct
      description: Removes a product from active management while preserving immutable version and audit history. Products with user application history must be retired instead.
      security: [{bearerAuth: []}]
      parameters:
        - {name: productID, in: path, required: true, schema: {type: string, format: uuid}}
      responses:
        '200': {description: Product removed from the active catalog}
        '404': {description: Product not found}
        '409': {description: Product has application history or an active version test}
    patch:
      operationId: updateProduct
      description: Updates the display name while preserving the product slug and all immutable versions.
      security: [{bearerAuth: []}]
      parameters:
        - {name: productID, in: path, required: true, schema: {type: string, format: uuid}}
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name: {type: string, minLength: 1, maxLength: 120}
      responses:
        '200': {description: Product name updated idempotently and audited when changed}
        '404': {description: Product not found}
...
components:
  securitySchemes:
    bearerAuth: {type: http, scheme: bearer}
  schemas:
...
    ProductDependency:
      type: object
      additionalProperties: false
      required: [key, productId, serviceSlug, required]
      properties:
        key: {type: string, pattern: '^[a-z][a-z0-9-]{0,31}$'}
        productId: {type: string, format: uuid}
        serviceSlug: {type: string, pattern: '^[a-z0-9][a-z0-9-]{0,62}$'}
        required: {type: boolean}
...